Prevent from adding product to cart more than once. Fixed error while deleting product from cart if it's not there. Added patch endpoint to carts controller. Prevent from creating multiple users with the same email. Fixed error while signing in on unexisting account.

+++ Szop/Controllers/CartsController.cs
@@ -2,6 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.Data.Entity;
 using System.Linq;
+using System.Net;
+using System.Net.Http;
 using System.Threading;
 using System.Web;
 using System.Web.Http;
@@ -19,21 +21,25 @@ namespace Szop.Controllers
 
         //get add delete buy
         // POST <controller>
-        public int Add([FromBody]DBCart value)
+        public HttpResponseMessage Add([FromBody]DBCart value)
         {
             var authenticationIdentity = Thread.CurrentPrincipal.Identity as JWTAuthenticationIdentity;
-            int UserId = authenticationIdentity.UserId;
+            int userId = authenticationIdentity.UserId;
+
+            if (db.Carts.Any(c => c.UserId == userId && c.ProductId == value.ProductId))
+                return Request.CreateResponse(HttpStatusCode.Forbidden, $"Product with id {value.ProductId} is already in the cart.", Configuration.Formatters.JsonFormatter);
+
             DBCart toAdd = new DBCart
             {
                 Id = value.Id,
-                UserId = UserId,
+                UserId = userId,
                 ProductId = value.ProductId,
                 Quantity = value.Quantity
             };
-            DBCart added = db.Carts.Add(toAdd);
+            db.Carts.Add(toAdd);
             db.SaveChanges();
 
-            return added.Id;
+            return Request.CreateResponse(HttpStatusCode.OK);
         }
 
         // GET <controller>
@@ -47,25 +53,49 @@ namespace Szop.Controllers
 
         // DELETE carts/{productId}
         [HttpDelete]
-        public IHttpActionResult Delete(int id)
+        public HttpResponseMessage Delete(int id)
         {
             var authenticationIdentity = Thread.CurrentPrincipal.Identity as JWTAuthenticationIdentity;
             int userId = authenticationIdentity.UserId;
 
-            DBCart cart = db.Carts.Where(c =>
+            IEnumerable<DBCart> crds = db.Carts.Where(c =>
                 c.UserId == userId &&
                 c.ProductId == id
-            ).First();
+            );
+
+            if (crds.Count() == 0)
+                return Request.CreateResponse(HttpStatusCode.NotFound, $"Product with id {id} doesn't exist.", Configuration.Formatters.JsonFormatter);
+
+            DBCart cart = crds.First();
 
             if (cart == null)
-            {
-                return NotFound();
-            }
+                return Request.CreateResponse(HttpStatusCode.NotFound, $"Product with id {id} doesn't exist.", Configuration.Formatters.JsonFormatter);
 
             db.Carts.Remove(cart);
             db.SaveChanges();
 
-            return Ok(cart);
+            return Request.CreateResponse(HttpStatusCode.OK, cart, Configuration.Formatters.JsonFormatter);
+        }
+
+        [HttpPatch]
+        public HttpResponseMessage Patch([FromBody]DBCart value)
+        {
+            var authenticationIdentity = Thread.CurrentPrincipal.Identity as JWTAuthenticationIdentity;
+            int userId = authenticationIdentity.UserId;
+
+            IEnumerable<DBCart> crds = db.Carts.Where(c =>
+                c.UserId == userId &&
+                c.ProductId == value.ProductId
+            );
+
+            if (crds.Count() == 0)
+                return Request.CreateResponse(HttpStatusCode.Forbidden, $"Product with id {value.ProductId} is not in the cart.", Configuration.Formatters.JsonFormatter);
+
+            DBCart cart = crds.First();
+            cart.Quantity = value.Quantity;
+            db.SaveChanges();
+
+            return Request.CreateResponse(HttpStatusCode.OK);
         }
 
         internal Cart Map(DBCart dbCart)
